requirements.txt
Django==1.7.1
django-polymorphic==0.6
